requirements.txt
click==6.7
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.13
imutils==0.4.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
opencv-contrib-python==3.3.0.9
Pillow==5.1.0
pkg-resources==0.0.0
Werkzeug==0.14.1
